:recycle: Extract Search component
