+++ src/pages/posts/index.tsx
@@ -1,4 +1,3 @@
-import { useEffect, useRef, useState } from 'react';
 import Image from 'next/image';
 import Card from '@/components/Card';
 import Date from '@/components/Date';
@@ -8,6 +7,8 @@ import { getSortedContentMetadata } from '@/lib/content';
 import { Post, PostAttributes } from '@/lib/models/content';
 import Catalog from '@/components/Catalog';
 import { InferGetStaticPropsType } from 'next';
+import { ChangeEvent, useState } from 'react';
+import Search from '@/components/Search';
 
 export async function getStaticProps() {
   const allPostsData = getSortedContentMetadata<PostAttributes, Post>('posts');
@@ -22,8 +23,10 @@ export default function Posts({
   allPostsData,
 }: InferGetStaticPropsType<typeof getStaticProps>) {
   const [search, setSearch] = useState('');
-  const inputRef = useRef<HTMLInputElement>(null);
   const showTags = false;
+
+  const handleSearch = (event: ChangeEvent<HTMLInputElement>) =>
+    setSearch(event.target.value);
 
   const posts = allPostsData.filter((post) => {
     const {
@@ -32,18 +35,6 @@ export default function Posts({
     const haystack = `${title} ${tags.join(' ')} ${description}`;
     return haystack.toLowerCase().includes(search.toLowerCase());
   });
-
-  const handleKeydown = (event: KeyboardEvent) => {
-    if (event.key === '/' && document.activeElement !== inputRef.current) {
-      event.preventDefault();
-      inputRef.current?.focus();
-    }
-  };
-
-  useEffect(() => {
-    window.addEventListener('keydown', handleKeydown);
-    return () => window.removeEventListener('keydown', handleKeydown);
-  }, []);
 
   return (
     <Layout
@@ -57,21 +48,7 @@ export default function Posts({
             </h1>
           }
         >
-          <div className="flex">
-            <input
-              className="flex-grow w-full px-4 py-2 bg-gray-300 border-2 border-gray-400 rounded sm:border-r-0 sm:rounded-r-none border-opacity-20 focus:outline-none focus:bg-opacity-30 hover:bg-opacity-30 bg-opacity-10"
-              placeholder="Type what you're looking for..."
-              type="text"
-              onChange={(event) => setSearch(event.target.value)}
-              ref={inputRef}
-            />
-            <kbd
-              className="hidden h-full p-2 my-auto text-gray-400 bg-gray-400 border-2 border-gray-400 rounded rounded-l-none border-opacity-20 bg-opacity-20 sm:block"
-              title="Press / to focus"
-            >
-              /
-            </kbd>
-          </div>
+          <Search onSearch={handleSearch} />
         </Card>
       </div>
       {posts.length ? (
